requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Django==2.2.7
django-humanize==0.1.2
django-taggit==1.2.0
django-tinymce==2.8.0
djangorestframework==3.11.0
humanize==0.5.1
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.0
Pillow==6.2.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7